feat(validation): 로그인 검증 로직 분리 및 제네릭 타입 변경
- 공용화를 위해 T 타입으로 변경
