+++ src/utils/validation.ts
@@ -1,53 +1,28 @@
 import type { Dispatch, SetStateAction } from "react";
-import { EMAIL_REGEX, ERROR_MESSAGE } from "@/constants";
-import type { FormData, ValidationErrors, ValidationRule } from "@/utils/type";
+import type { ValidationErrors, ValidationRule } from "@/utils/type";
 
-export const isValidEmail = (email: string): boolean => EMAIL_REGEX.test(email);
-
-export const validateField = (
-  value: string,
-  rules: ValidationRule[],
+export const validateField = <T>(
+  value: T,
+  rules: ValidationRule<T>[],
 ): string | undefined => {
   const failedRule = rules.find(rule => rule.condition(value));
   return failedRule ? failedRule.message : undefined;
 };
 
-export const validators = {
-  id: (value: string) =>
-    validateField(value, [
-      { condition: val => !val.trim(), message: ERROR_MESSAGE.ID.REQUIRED },
-      {
-        condition: val => !isValidEmail(val),
-        message: ERROR_MESSAGE.ID.INVALID_FORMAT,
-      },
-    ]),
-
-  password: (value: string) =>
-    validateField(value, [
-      {
-        condition: val => !val.trim(),
-        message: ERROR_MESSAGE.PASSWORD.REQUIRED,
-      },
-      {
-        condition: val => val.length < 8,
-        message: ERROR_MESSAGE.PASSWORD.MIN_LENGTH,
-      },
-    ]),
-};
-
-export const createFieldHandler = <T extends keyof ValidationErrors>(
-  field: T,
-  setFormData: Dispatch<SetStateAction<FormData>>,
-  setErrors: Dispatch<SetStateAction<ValidationErrors>>,
+export const createFieldHandler = <T, K extends keyof T>(
+  field: K,
+  validators: Record<K, (value: T[K]) => string | undefined>,
+  setFormData: Dispatch<SetStateAction<T>>,
+  setErrors: Dispatch<SetStateAction<ValidationErrors<T>>>,
   hasError: boolean,
 ) => ({
-  onChange: (value: string) => {
+  onChange: (value: T[K]) => {
     setFormData(prev => ({ ...prev, [field]: value }));
     if (hasError) {
       setErrors(prev => ({ ...prev, [field]: undefined }));
     }
   },
-  onBlur: (value: string) => {
+  onBlur: (value: T[K]) => {
     const error = validators[field](value);
     setErrors(prev => ({ ...prev, [field]: error }));
   },
